requirements.txt
altair==5.0.1
appdirs==1.4.4
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
cmake==3.25.0
coloredlogs==15.0.1
contourpy==1.1.0
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.0
decorator==5.1.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lit==15.0.7
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
nvidia-tensorrt==99.0.0
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
platformdirs==3.10.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pycuda==2022.2.2
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytools==2023.1.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
requests==2.28.1
rich==13.5.2
rpds-py==0.9.2
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
sympy==1.11.1
tenacity==8.2.2
tensorrt==8.6.1
tensorrt-bindings==8.6.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.147
urllib3==1.26.13
validators==0.20.0
watchdog==3.0.0
zipp==3.16.2
